backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
psycopg[binary]==3.2.3  # Modern PostgreSQL adapter (psycopg3)
sqlalchemy==2.0.35
alembic==1.13.3

# Redis & Caching
redis==5.1.1
hiredis==3.0.0

# Background Tasks
celery[redis]==5.4.0  # Async task queue with Redis support
flower==2.0.1  # Celery monitoring UI

# HTTP Client
httpx==0.27.2
aiohttp==3.10.10

# Data Validation & Serialization
python-multipart==0.0.12
email-validator==2.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Monitoring & Logging
sentry-sdk[fastapi]==2.15.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-httpx==0.33.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.11.2
pre-commit==4.0.1

# AI/ML (will be expanded in Task 2)
openai==1.52.0  # For OpenAI-compatible APIs including Cerebras
anthropic==0.39.0  # Claude streaming and batch API support

# Firebase SDK
firebase-admin==6.5.0  # Firebase Admin SDK for server-side operations
google-cloud-firestore==2.19.0  # Firestore client library
google-cloud-storage==2.18.2  # Firebase Storage operations

# Document Processing
pdfplumber==0.11.4  # Best PDF text extraction (Context7 recommended)
python-docx==1.1.2  # DOCX document parsing
pypandoc==1.13  # Universal document conversion

# Social Media Scraping
tweepy==4.14.0  # Twitter API v2 client
praw==7.7.1  # Reddit API wrapper

# Vector Embeddings
sentence-transformers==3.3.1  # For generating text embeddings
numpy==2.0.2  # Required for vector operations
pgvector==0.3.6  # PostgreSQL vector extension support
